Cleanup, moved exisitng tests and support frameworks to malcolm category, now we know how we're going to organize this. They won't be touched for a while.
