+++ testutils/malcolm/test_acquireperiod.py
