+++ testutils/malcolm/test_acquirereadout.py
