+++ testutils/malcolm/test_imageacquire.py
