+++ testutils/malcolm/test_imagemode.py
