+++ testutils/malcolm/test_numimages.py
