+++ testutils/malcolm/test_triggermode.py
